requirements.txt
certifi==2018.10.15
chardet==3.0.4
idna==2.7
praw==6.0.0
prawcore==1.0.0
python-dotenv==0.9.1
requests==2.20.0
update-checker==0.16
urllib3==1.24
2